More sensible component tree
